Improved file structure
