--- cmd/server/config/config.go
@@ -1,33 +1,0 @@
-package config
-
-import (
-	"encoding/json"
-	"os"
-)
-
-type Server struct {
-	URL    string
-	Weight int
-}
-
-type ServerConfig struct {
-	Servers []Server `json:"servers"`
-}
-
-func LoadConfig() ([]Server, error) {
-	f, err := os.Open("cmd/server/config/config.json")
-
-	if err != nil {
-		return nil, err
-	}
-
-	defer f.Close()
-
-	var cfg ServerConfig
-
-	if err := json.NewDecoder(f).Decode(&cfg); err != nil {
-		return nil, err
-	}
-
-	return cfg.Servers, nil
-}
