+++ internal/config.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"encoding/json"
 	"errors"
 	"os"
 
@@ -26,3 +27,31 @@ func ParseEnv() (*Config, error) {
 		Port: port,
 	}, nil
 }
+
+type Server struct {
+	URL    string
+	Weight int
+}
+
+type ServerConfig struct {
+	Servers []Server `json:"servers"`
+}
+
+func LoadConfig() ([]Server, error) {
+	f, err := os.Open("weighted_robin.config.json")
+
+	if err != nil {
+		return nil, err
+	}
+
+	defer f.Close()
+
+	var cfg ServerConfig
+
+	if err := json.NewDecoder(f).Decode(&cfg); err != nil {
+		return nil, err
+	}
+
+	return cfg.Servers, nil
+}
+
